Massive updates/cleanups to the middleware API
This commit converts all the API to use the load() function to get POST
variables, and the $this->memberid value to enforce authentication.

The result of mysql_query only returns false on an error. It DOES NOT
return false if the result is an empty set:

	$res = mysql_query("stuff");
	if(!res)
		do_shit;

If "stuff" returns no results, do_shit will not be called. You
have to do:

	if(!mysql_num_rows($res))

...if you want that effect.