Combine all getImages() variants into one API function
This commit condenses 208 lines of code into 42. All the previous
methods of getting images by various criteria are now submitted to this
singe function, which processes the request based on the variables
passed to it. Images are output in the natural ordering of the primary
key in the database, and you can now specify an OFFSET variable, which
will skip that many images in the natural ordering (allowing for album
pagination, etc.). NUM will limit the number of results returned. The
API currently does not force a NUM value - we may wish to do this in the
future.